requirements.txt
Authlib==1.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
decorator==5.1.1
idna==3.4
pycparser==2.21
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.3
validators==0.20.0
weaviate-client==3.21.0
rfc3339~=6.2
pandas~=2.0.2